Remove recursive serialization
